stage_1/requirements.txt
absl-py==2.1.0
accelerate==0.34.0
addict==2.4.0
aeon==0.9.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
braceexpand==0.1.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
comm==0.2.2
confection==0.1.5
ConfigArgParse==1.7
contexttimer==0.3.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
decord==0.6.0
Deprecated==1.2.18
diffusers==0.14.0
distlib==0.3.8
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
executing==2.1.0
fairscale==0.4.4
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
h5py==3.11.0
huggingface-hub==0.24.6
identify==2.6.0
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
immutabledict==4.2.0
importlib_metadata==8.4.0
importlib_resources==6.4.4
iopath==0.1.10
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaggle==1.6.17
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
lightning==2.3.3
lightning-utilities==0.11.7
llvmlite==0.41.1
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
narwhals==1.6.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
nodeenv==1.9.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open3d==0.18.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
ortools==9.11.4210
packaging==24.1
pandas==2.0.3
parso==0.8.4
peft==0.12.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.24.0
portalocker==2.10.1
pre-commit==3.5.0
preshed==3.0.9
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==5.26.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-slugify==8.0.4
pytorch-lightning==2.4.0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
retrying==1.3.4
rich==13.8.0
rpds-py==0.20.0
safetensors==0.4.4
salesforce-lavis==1.0.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
sentry-sdk==2.13.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.38.0
sympy==1.13.2
tenacity==8.5.0
text-unidecode==1.3
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
torch==2.4.0
torchaudio==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.26.1
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
wandb==0.17.8
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.100
webencodings==0.5.1
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.16.0
yarl==1.9.10
zipp==3.20.1
